fixed bug

+++ src/Lib/Utls.ts
@@ -1,7 +1,7 @@
 export let GetSymbolValue = function (symbolStr: string) {
     let targetArr = symbolStr.split(" ");
     return {
-        Value: parseInt(targetArr[0]),
+        Value: parseFloat(targetArr[0]),
         Symbol: targetArr[1]
     }
 }
